Initial implementation of reverse-highlight
Matches in the opposite direction are highlighted in dark magenta.
Seems to be working as expecting, but still needs:

* settings to disable/enable or change the hl color
* further manual testing
* unit tests
* documentation

When I run the test suite in Vader, I'm seeing many failures. However, I
also see errors when I run the tests on the master branch, and the
number of failures is inconsistent.